Revert "Simplify dev resource path detection"
This reverts commit ba12a1e2fc12543ff49d1c9add6b997086574961.
