+++ src/main-process/main.js
@@ -5,10 +5,8 @@ if (typeof snapshotResult !== 'undefined') {
 const startTime = Date.now()
 
 const path = require('path')
-const fs = require('fs-plus')
-const CSON = require('season')
 const yargs = require('yargs')
-const electron = require('electron')
+const getDevResourcePath = require('./get-dev-resource-path')
 
 const args =
   yargs(process.argv)
@@ -16,34 +14,14 @@ const args =
     .alias('t', 'test')
     .argv
 
-function isAtomRepoPath (repoPath) {
-  let packageJsonPath = path.join(repoPath, 'package.json')
-  if (fs.statSyncNoException(packageJsonPath)) {
-    let packageJson = CSON.readFileSync(packageJsonPath)
-    return packageJson.name === 'atom'
-  }
-
-  return false
-}
-
 let resourcePath
 
 if (args.resourcePath) {
   resourcePath = args.resourcePath
 } else {
   const stableResourcePath = path.dirname(path.dirname(__dirname))
-  const defaultRepositoryPath = path.join(electron.app.getPath('home'), 'github', 'atom')
-
   if (args.dev || args.test || args.benchmark || args.benchmarkTest) {
-    if (process.env.ATOM_DEV_RESOURCE_PATH) {
-      resourcePath = process.env.ATOM_DEV_RESOURCE_PATH
-    } else if (isAtomRepoPath(process.cwd())) {
-      resourcePath = process.cwd()
-    } else if (fs.statSyncNoException(defaultRepositoryPath)) {
-      resourcePath = defaultRepositoryPath
-    } else {
-      resourcePath = stableResourcePath
-    }
+    resourcePath = getDevResourcePath() || stableResourcePath
   } else {
     resourcePath = stableResourcePath
   }
