+++ src/main-process/parse-command-line.js
@@ -5,8 +5,9 @@ const yargs = require('yargs')
 const {app} = require('electron')
 const path = require('path')
 const fs = require('fs-plus')
+const getDevResourcePath = require('./get-dev-resource-path')
 
-module.exports = function parseCommandLine (processArgs, initialResourcePath) {
+module.exports = function parseCommandLine (processArgs) {
   const options = yargs(processArgs).wrap(yargs.terminalWidth())
   const version = app.getVersion()
   options.usage(
@@ -119,7 +120,7 @@ module.exports = function parseCommandLine (processArgs, initialResourcePath) {
   let pathsToOpen = []
   let urlsToOpen = []
   let devMode = args['dev']
-  let devResourcePath = initialResourcePath
+  let devResourcePath = getDevResourcePath()
   let resourcePath = null
 
   for (const path of args._) {
